Move _find_run_command from template to framework
